habit-tracker/requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
WTForms==3.2.1
python-dateutil==2.9.0
Werkzeug==3.1.3
psycopg2-binary==2.9.10
pytest==8.4.1
pytest-cov==5.0.0
flake8==7.3.0
python-dotenv==1.1.1
gunicorn==22.0.0